add user verification to add, delete and edit ads endpoints

+++ controllers/ads.controller.js
@@ -9,7 +9,7 @@ exports.getAll = async (req, res) => {
   try {
     res.json(await Ad.find());
   }
-  catch(err) {
+  catch (err) {
     res.status(500).json({ message: err });
   }
 };
@@ -20,10 +20,10 @@ exports.getById = async (req, res) => {
 
   try {
     const adToSelect = await Ad.findById(req.params.id);
-    if(!adToSelect) res.status(404).json({ message: 'Ad not found' });
+    if (!adToSelect) res.status(404).json({ message: 'Ad not found' });
     else res.json(adToSelect);
   }
-  catch(err) {
+  catch (err) {
     res.status(500).json({ message: err });
   }
 
@@ -31,73 +31,76 @@ exports.getById = async (req, res) => {
 
 
 exports.addNew = async (req, res) => {
-
   try {
-  const { title, content, date, price, adress, seller } = req.body;
-  const fileType = req.file ? await getImageFileType(req.file) : 'unknown';
+    const { title, content, date, price, address } = req.body;
+    const fileType = req.file ? await getImageFileType(req.file) : 'unknown';
 
-  if (title && typeof title === 'string' && content && typeof content === 'string' && date && 
-  price && adress && seller && req.file && ['image/png', 'image/jpeg', 'image/gif'].includes(fileType)) {
+    if (
+      title &&
+      typeof title === 'string' &&
+      content &&
+      typeof content === 'string' &&
+      date &&
+      price &&
+      address &&
+      req.file &&
+      ['image/png', 'image/jpeg', 'image/gif'].includes(fileType)
+    ) {
+      const newAd = new Ad({
+        title,
+        content,
+        date,
+        photo: req.file.filename,
+        price,
+        address,
+        seller: req.session.login,
+      });
 
-
-  const newAdd = new Ad({ title, content, date, photo: req.file.filename, price, adress, seller });
-    await newAdd.save();
-    res.json({ message: 'OK' });
-
-  } else {
-    if (req.file) {
-      fs.unlinkSync('./public/uploads/' + req.file.filename);
-  }
-  res.status(400).send({ message: 'Bad request' });
-  }
-
-  } catch(err) {
+      await newAd.save();
+      res.json({ message: 'OK' });
+    } else {
+      if (req.file) {
+        fs.unlinkSync('./public/uploads/' + req.file.filename);
+      }
+      res.status(400).send({ message: 'Bad request' });
+    }
+  } catch (err) {
     res.status(500).json({ message: err });
   }
 };
 
-// exports.edit = async (req, res) => {
-//   try {
-//       const { title, content, date, photo, price, adress, seller } = req.body;
-
-  
-//     const adToUpdate = await Ad.findById(req.params.id);
-//     if(adToUpdate) {
-//       await Ad.updateOne({ _id: req.params.id }, { $set: { title, content, date, photo, price, adress, seller }});
-//       res.status(200).json({ message: 'OK' });
-//     }
-//    else res.status(404).json({ error: 'Add not found' });
-//   }
-//   catch(err) {
-//     res.status(500).json({ message: err });
-//   }
-
-// };
-
 exports.edit = async (req, res) => {
   try {
-    const { title, content, date, price, adress, seller } = req.body;
+    const { title, content, date, price, address } = req.body;
     const adToUpdate = await Ad.findById(req.params.id);
 
     if (!adToUpdate) {
       return res.status(404).json({ error: 'Ad not found' });
     }
 
+
+    if (req.session.login !== adToUpdate.seller.toString()) {
+      return res.status(403).json({ error: 'Forbidden - Not authorized to edit this ad' });
+    }
+
     if (req.file) {
-
-      if (adToUpdate.photo) {
-        fs.unlinkSync('./public/uploads/' + adToUpdate.photo);
+      const fileType = await getImageFileType(req.file);
+      if (req.file && ['image/png', 'image/jpeg', 'image/gif'].includes(fileType)) {
+        if (adToUpdate.photo) {
+          fs.unlinkSync('./public/uploads/' + adToUpdate.photo);
+        }
+      } else {
+        res.status(400).send({ message: 'Bad request' });
       }
-
       adToUpdate.photo = req.file.filename;
     }
+
 
     adToUpdate.title = title || adToUpdate.title;
     adToUpdate.content = content || adToUpdate.content;
     adToUpdate.date = date || adToUpdate.date;
     adToUpdate.price = price || adToUpdate.price;
-    adToUpdate.adress = adress || adToUpdate.adress;
-    adToUpdate.seller = seller || adToUpdate.seller;
+    adToUpdate.address = address || adToUpdate.address;
 
     await adToUpdate.save();
 
@@ -106,21 +109,27 @@ exports.edit = async (req, res) => {
     res.status(500).json({ message: err });
   }
 };
-
 exports.delete = async (req, res) => {
-
   try {
     const adToDelete = await Ad.findById(req.params.id);
-    if(adToDelete) {
+    if (adToDelete) {
+
+      if (req.session.login !== adToDelete.seller.toString()) {
+        return res.status(403).json({ error: 'Forbidden - Not authorized to delete this ad' });
+      }
+
+      if (adToDelete.photo) {
+        fs.unlinkSync(`./public/uploads/${adToDelete.photo}`);
+      }
+
       await Ad.deleteOne({ _id: req.params.id });
       res.status(200).json({ message: 'Ad deleted successfully' });
+    } else {
+      res.status(404).json({ message: 'Not found...' });
     }
-    else res.status(404).json({ message: 'Not found...' });
-  }
-  catch(err) {
+  } catch (err) {
     res.status(500).json({ message: err });
   }
-
 };
 
 exports.searchAds = async (req, res) => {
